first test

+++ memecity.engine/Engine/SDL/TimerFacade.h
@@ -9,7 +9,7 @@ namespace memecity::engine::sdl {
 	private:
 		unsigned int start_ticks{};
 		unsigned int elapsed_ticks{};
-		float delta_time{};
+		double delta_time{};
 
 	public:
 		TimerFacade();
@@ -18,7 +18,7 @@ namespace memecity::engine::sdl {
 		void reset();
 
 		///<summary>Returns the current delta time.</summary>
-		float get_delta_time() const;
+		double get_delta_time() const;
 
 		///<summary>Sets delta time to elapsed time since last update.</summary>
 		void update();
